# Conflicts:
#	core/src/dev/adamhodgkinson/game/MeleeWeapon.java

+++ core/src/dev/adamhodgkinson/WeaponData.java
@@ -1,9 +1,9 @@
 package dev.adamhodgkinson;
 
 public class WeaponData {
-    public int damage;
-    public int range;
-    public int attackspeed; // time in ms between hits, higher value is slower
+    public float damage;
+    public float range;
+    public float attackspeed; // time in ms between hits, higher value is slower
     public float knockback;
     public String textureName;
     public boolean isMelee;
